+++ src/main/java/employee/service/EmployeeService.java
@@ -4,9 +4,9 @@ import employee.data.EmployeeJPA;
 
 public interface EmployeeService {
 
-    public Object getEmployee(String id, double version);
-    public Object addEmployee(EmployeeJPA employee, double version);
-    public Object updateEmployee(String id, EmployeeJPA employee, double version);
-    public Object deleteEmployee(String id, double version);
+    public Object getEmployee(String id, int version);
+    public Object addEmployee(EmployeeJPA employee, int version);
+    public Object updateEmployee(String id, EmployeeJPA employee, int version);
+    public Object deleteEmployee(String id, int version);
 
 }
